Optional methods and fixes
If you have a file called options with the word True in it in the same
folder as the program it will generate two extra methods to set the
rotation of the structure and get the number of available rotations. The
program no longer sets non solid blocks multiple times in methods with
rotation.